+++ Gemfile
@@ -88,3 +88,5 @@ gem 'active_model_serializers', '~> 0.10.13'
 gem 'bcrypt', '~> 3.1', '>= 3.1.19'
 
 gem 'jwt'
+
+gem 'bcrypt-ruby'
