+++ app/controllers/api/v1/posts_controller.rb
@@ -1,11 +1,6 @@
 class Api::V1::PostsController < ApplicationController
   def index
     @posts = Post.includes(comments: [:author]).where(author_id: params[:user_id])
-
-    if @posts
-      render json: @posts.to_json, status: :ok
-    else
-      render json: { status: 'Error', errors: @posts.errors }
-    end
+    render json: @posts, status: :ok
   end
 end
