+++ app/helpers/authentication_helper.rb
@@ -1,0 +1,2 @@
+module AuthenticationHelper
+end
