+++ config/routes.rb
@@ -14,10 +14,20 @@ Rails.application.routes.draw do
 
   namespace :api do
     namespace :v1 do
-      get '/users/:user_id/posts', to: 'posts#index'
-      get '/users/:user_id/posts/:post_id/comments', to: 'comments#index'
-      post '/users/:user_id/posts/:post_id/comments', to: 'comments#create'
+      resources :users do
+        resources :posts, only: [:index] do
+          resources :comments, only: [:index, :create]
+        end
+      end
     end
   end
+
+  # namespace :api do
+  #   namespace :v1 do
+  #     get '/users/:user_id/posts', to: 'posts#index'
+  #     get '/users/:user_id/posts/:post_id/comments', to: 'comments#index'
+  #     post '/users/:user_id/posts/:post_id/comments', to: 'comments#create'
+  #   end
+  # end
 
 end
